requirements.txt
fastapi
uvicorn
firebase-admin
stripe
python-dotenv
cryptography